module/requirements_module.txt
PyYAML==6.0.1
picamera2==0.3.18